* int로 자동 형변환 불가능한 플랫폼을 위해 Wid 를 int로 명시적 형변환

+++ src/core/nativerenderer.cpp
@@ -66,7 +66,7 @@ void NativeRenderer::setFrameInfo(const VideoFrame::Info &frame) {
 }
 
 int NativeRenderer::screenWinId() const {
-	return d->screen->winId();
+	return (int)d->screen->winId();
 }
 
 void NativeRenderer::rerender() {
